prediction_saver/model_service/requirements.txt
fastapi
uvicorn[standard]
pyspark
joblib
pandas
requests
redis
pydantic
kafka-python